pyproject.toml
[tool.poetry]
name = "async-tkinter-loop"
version = "0.3.0"
description = "Asynchronous mainloop implementation for tkinter"
authors = ["insolor <insolor@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/insolor/async-tkinter-loop"
repository = "https://github.com/insolor/async-tkinter-loop"
keywords = ["tkinter", "async", "async-await", "asyncio", "aiohttp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.7"
asyncio = "^3.4.3"
Pillow = {version = "^9.1.1", optional = true}
httpx = {version = "^0.23.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-timeout = "^2.0.1"
flake8 = "^4.0.1"
tkinter-stubs = { url = "https://github.com/insolor/tkinter-stubs/releases/download/0.1.0/tkinter_stubs-0.1.0-py3-none-any.whl" }
black = "^22.3.0"

[tool.poetry.extras]
examples = ["httpx", "Pillow"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
